testing IPC for double

+++ src/RLFuzzing.py
@@ -16,7 +16,7 @@ while True:
     except sysv_ipc.ExistentialError:
         print("ERROR: message queue creation failed")
 
-    msg_npy = np.arange(BUFF_SIZE_SENDER, dtype=np.uint8).reshape((2,BUFF_SIZE_SENDER//2))
+    msg_npy = np.arange(BUFF_SIZE_SENDER, dtype=np.double).reshape((2,BUFF_SIZE_SENDER//2))
 
     try:
         mq_sender = sysv_ipc.MessageQueue(2, sysv_ipc.IPC_CREAT)
